Create DayTwoSolutionPt2.cpp
Solution to Part 2 of Day 2. This code is basically the part 1 solution, but modified to accomplish the task of Part 2 as well. This means forgoing a nested while loop as we now want to process all of the characters.

In Part 1, we wanted to discover a failure case as early as possible and abort the inner loop so that the outer loop could just walk through the characters until the next new line. At that point, we refresh the inner loop and rinse/repeat.

In Part 2, we need to look at every round played in each game in order to determine the max of each color of cube as this is our minimum number of cubes necessary to make the game true. This means processing all of the input, rather than skipping input that comes after a failure case (Part 1). But we still want to accomplish Part 1 since we are taking the time to walk through it, so we do. Hence the modifications.