Rename DayOneSolution.cpp to DayOneSolutionPt1.cpp
Renamed: I didn't know there was a Part 2, lol.

As for the solution itself, it's a switch statement that filters out the calibration data. The first character digit encounter is converted to its appropriate integer as dictated by the switch case and stored into the tens place (first index) of our array. The placeFlag is incremented, telling us that subsequent numbers are to be filtered into the ones place (second element).

Once a line has had its numbers extracted, a bit of math on our two element array gets the proper two-digit number for that line. Rinse and repeat.

Note: the switch is mindless and so it will filter out and place all character digits (after the first) into the second array element. A bit of extra work, but we're working at the character level and it keeps the logic simple.